+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [FIX] Upgrade google appcompanist
 - [FIX] Bump deps, remove PR apks and fix wearos track
 - [FIX] Return empty list, if flipper path not exist
+- [FIX] New Bottom Sheet
 - [FIX] Wearable app
 
 # 1.6.3